requirements.txt
confluent_kafka==2.3.0
kafka-python==2.0.2
psycopg2-binary==2.9.9
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
requests==2.31.0